requirements.txt
Django>=4.2.4
gunicorn>=21.2.0
psycopg2-binary>=2.9.7
django-taggit>=4.0.0
djangorestframework>=3.14.0
django-filter>=23.2
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.26.4
drf-spectacular-sidecar>=2023.9.1
boto3>=1.28.44
django-storages>=1.14
Pillow>=10.0.0
